piscina cliente fatto

+++ gestione/cliente/abbonamenti/view_abbonamenti/view_inserisci_abbonamenti.py
@@ -203,7 +203,7 @@ class view_inserisci_abbonamenti(QWidget):
             return
 
         try:
-            cellulare = float(self.campo_cellulare.text())
+            cellulare = int(self.campo_cellulare.text())
 
         except:
